[fe] refactor: 팀이름 선택시가져오는것을 mock데이터에서 가져오는것으로 변경
