+++ FE/baseball/src/components/GameProgression/Defense.jsx
@@ -18,8 +18,10 @@ function Defense({ location }) {
   const { dispatch } = useContext(BaseballContext);
 
   useEffect(() => {
-    const { teamType, payload } = location.teamInfo;
-    dispatch({ type: SET_TEAM_NAME, teamType, payload });
+    if (!location.teamInfo) return;
+
+    const { payload } = location.teamInfo;
+    dispatch({ type: SET_TEAM_NAME, payload });
   }, [location]);
 
   return (
